增加User-Agent过滤模式
代码清理
屏蔽日志上报
屏蔽更新检查
屏蔽初始化弹框跳转

+++ MyHelper/SelfUpgradeHelp.cs
@@ -1,24 +1,22 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Net.Http;
-using System.Text;
+using System.Reflection;
 using System.Threading.Tasks;
-using System.Windows.Forms;
 
 namespace FreeHttp.MyHelper
 {
     public class SelfUpgradeHelp
     {
-        private static HttpClient httpClient = new HttpClient();
+        private static readonly HttpClient httpClient = new HttpClient();
+
         private static string GetFreeHttpDllPath()
         {
             string path = null;
             try
             {
-                string codeBase = System.Reflection.Assembly.GetExecutingAssembly().CodeBase;
-                UriBuilder uri = new UriBuilder(codeBase);
+                var codeBase = Assembly.GetExecutingAssembly().CodeBase;
+                var uri = new UriBuilder(codeBase);
                 path = Uri.UnescapeDataString(uri.Path);
                 //Path.GetDirectoryName(path);
             }
@@ -28,17 +26,15 @@ namespace FreeHttp.MyHelper
             }
             finally
             {
-                if(string.IsNullOrEmpty(path))
-                {
-                    path = Directory.GetCurrentDirectory() + "\\Scripts\\FreeHttp.dll";
-                }
+                if (string.IsNullOrEmpty(path)) path = Directory.GetCurrentDirectory() + "\\Scripts\\FreeHttp.dll";
             }
+
             return path;
         }
+
         private static async Task<bool> DownloadUpgradeFileAsync(string uri, string path)
         {
             if (File.Exists(path))
-            {
                 try
                 {
                     File.Delete(path);
@@ -47,12 +43,12 @@ namespace FreeHttp.MyHelper
                 {
                     return false;
                 }
-            }
-            HttpResponseMessage response = await httpClient.GetAsync(uri);
+
+            var response = await httpClient.GetAsync(uri);
             if (response.IsSuccessStatusCode)
             {
-                Stream fileStrem = await response.Content.ReadAsStreamAsync();
-                using (FileStream fs = new FileStream(path, FileMode.Create, FileAccess.Write, FileShare.None))
+                var fileStrem = await response.Content.ReadAsStreamAsync();
+                using (var fs = new FileStream(path, FileMode.Create, FileAccess.Write, FileShare.None))
                 {
                     await fileStrem.CopyToAsync(fs);
                     fs.Close();
@@ -62,30 +58,29 @@ namespace FreeHttp.MyHelper
             {
                 return false;
             }
+
             return true;
         }
+
         public static async Task<string> UpdateDllAsync(string sourceFileUrl)
         {
-            string path = GetFreeHttpDllPath();
-            string oldFilePath = path + ".oldversion";
-            string upgradeFile = Path.GetDirectoryName(path) + "/FreeHttpUpgradeFile";
+            var path = GetFreeHttpDllPath();
+            var oldFilePath = path + ".oldversion";
+            var upgradeFile = Path.GetDirectoryName(path) + "/FreeHttpUpgradeFile";
             try
             {
-                if (File.Exists(oldFilePath))
-                {
-                    File.Delete(oldFilePath);
-                }
+                if (File.Exists(oldFilePath)) File.Delete(oldFilePath);
                 //https://lulianqi.com/file/FreeHttpUpgradeFile
                 await DownloadUpgradeFileAsync(sourceFileUrl, upgradeFile);
-                System.IO.File.Move(path, oldFilePath);
-                System.IO.File.Move(upgradeFile, path);
+                File.Move(path, oldFilePath);
+                File.Move(upgradeFile, path);
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 return ex.Message;
             }
+
             return null;
         }
-
     }
 }
